fix: address critical review issues on eBPF group_dead PR
- Read group_dead as u8 (not u32) in BPF kernel to match kernel's bool type
- Parse and validate tracepoint field size; disable group_dead if size unexpected
- Probe both /sys/kernel/tracing and /sys/kernel/debug/tracing tracefs mounts
- Rename booleans to has_exit_code_support/has_group_dead_support (project convention)
- Separate skipped probes from degraded capabilities in log messages
- Log actual errors from offset discovery instead of silently discarding via .ok()
- Add tests for parse_tracepoint_field with size extraction

+++ crates/logfwd-ebpf-proto/sensor-ebpf/sensor-ebpf-kern/src/main.rs
@@ -185,7 +185,9 @@ fn try_process_exit(ctx: &TracePointContext) -> Result<(), i64> {
             let offset = cfg.sched_process_exit_group_dead_offset as usize;
             // SAFETY: the offset is parsed from the kernel tracepoint format
             // before programs attach; read failures preserve the pid==tgid fallback.
-            if let Ok(group_dead) = unsafe { ctx.read_at::<u32>(offset) } {
+            // Read as u8: the kernel defines group_dead as bool (1 byte).
+            // Reading u32 would pull adjacent padding bytes, causing false positives.
+            if let Ok(group_dead) = unsafe { ctx.read_at::<u8>(offset) } {
                 should_emit = group_dead != 0;
             }
         }
